ClaimSyncAPI/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
azure-monitor-opentelemetry==1.6.4
bcrypt==4.1.2
PyJWT==2.8.0
azure-keyvault-secrets==4.8.0
azure-identity==1.17.1
azure-communication-email==1.0.0
azure-mgmt-appcontainers==3.1.0
azure-storage-blob==12.19.0
